+++ crates/biome_html_formatter/src/html/bogus/bogus_attribute.rs
@@ -1,0 +1,5 @@
+use crate::FormatBogusNodeRule;
+use biome_html_syntax::HtmlBogusAttribute;
+#[derive(Debug, Clone, Default)]
+pub(crate) struct FormatHtmlBogusAttribute;
+impl FormatBogusNodeRule<HtmlBogusAttribute> for FormatHtmlBogusAttribute {}