+++ crates/biome_js_analyze/src/lint/nursery.rs
@@ -14,6 +14,7 @@ pub mod no_label_without_control;
 pub mod no_misplaced_assertion;
 pub mod no_react_specific_props;
 pub mod no_restricted_imports;
+pub mod no_restricted_types;
 pub mod no_static_element_interactions;
 pub mod no_substr;
 pub mod no_undeclared_dependencies;
@@ -26,6 +27,7 @@ pub mod use_adjacent_overload_signatures;
 pub mod use_aria_props_supported_by_role;
 pub mod use_consistent_builtin_instantiation;
 pub mod use_consistent_curly_braces;
+pub mod use_consistent_member_accessibility;
 pub mod use_date_now;
 pub mod use_default_switch_clause;
 pub mod use_error_message;
@@ -59,6 +61,7 @@ declare_lint_group! {
             self :: no_misplaced_assertion :: NoMisplacedAssertion ,
             self :: no_react_specific_props :: NoReactSpecificProps ,
             self :: no_restricted_imports :: NoRestrictedImports ,
+            self :: no_restricted_types :: NoRestrictedTypes ,
             self :: no_static_element_interactions :: NoStaticElementInteractions ,
             self :: no_substr :: NoSubstr ,
             self :: no_undeclared_dependencies :: NoUndeclaredDependencies ,
@@ -71,6 +74,7 @@ declare_lint_group! {
             self :: use_aria_props_supported_by_role :: UseAriaPropsSupportedByRole ,
             self :: use_consistent_builtin_instantiation :: UseConsistentBuiltinInstantiation ,
             self :: use_consistent_curly_braces :: UseConsistentCurlyBraces ,
+            self :: use_consistent_member_accessibility :: UseConsistentMemberAccessibility ,
             self :: use_date_now :: UseDateNow ,
             self :: use_default_switch_clause :: UseDefaultSwitchClause ,
             self :: use_error_message :: UseErrorMessage ,