+++ crates/biome_js_analyze/tests/specs/suspicious/noFocusedTests/invalid.js
@@ -12,3 +12,15 @@ test["only"]("bar", function () {});
 
 fdescribe("foo", () => {});
 fit("foo", () => {});
+
+describe.only(bar, () => {});
+it.only(bar, () => {});
+test.only(bar, () => {});
+
+describe.only(foo.bar, () => {});
+it.only(foo.bar, () => {});
+test.only(foo.bar, () => {});
+
+describe.only(name = name || "bar", () => {});
+it.only(name = name || "bar", () => {});
+test.only(name = name || "bar", () => {});