+++ crates/biome_unicode_table/src/lib.rs
@@ -12,16 +12,27 @@ pub fn is_html_id_start(c: char) -> bool {
     ID_Start(c)
 }
 
-/// Tests if `c` is a valid start of a CSS identifier
+/// Is `c` a CSS non-ascii character.
 #[inline]
-pub fn is_css_id_start(c: char) -> bool {
-    c == '_' || c == '$' || ID_Start(c)
-}
-
-/// Tests if `c` is a valid continuation of a CSS identifier.
-#[inline]
-pub fn is_css_id_continue(c: char) -> bool {
-    c == '\u{200d}' || c == '\u{200c}' || ID_Continue(c)
+pub fn is_css_non_ascii(c: char) -> bool {
+    matches!(
+        c as u32,
+        0xB7
+        | 0xc0..=0xd6
+        | 0xd8..=0xf6
+        | 0xf8..=0x37D
+        | 0x37F..=0x1FFF
+        | 0x200C
+        | 0x200D
+        | 0x203F
+        | 0x2040
+        | 0x2070..=0x218F
+        | 0x2C00..=0x2FEF
+        | 0x3001..=0xD7FF
+        | 0xF900..0xFDCF
+        | 0xFDF0..0xFFFD
+        | 0x10000..
+    )
 }
 
 /// Tests if `c` is a valid start of a js identifier